fix(mobile): move login keyboard compensation to the shell
