+++ apps/mobile/src/components/login/email-otp-form.tsx
@@ -1,29 +1,11 @@
-import { useEffect, useRef, useState } from 'react';
-import {
-  ActivityIndicator,
-  AppState,
-  Keyboard,
-  type KeyboardEvent,
-  Platform,
-  TextInput,
-  View,
-} from 'react-native';
+import { useRef, useState } from 'react';
+import { ActivityIndicator, TextInput, View } from 'react-native';
 
 import { Button } from '@/components/ui/button';
 import { Text } from '@/components/ui/text';
-import {
-  resolveAppAwareKeyboardPadding,
-  resolveKeyboardPaddingEventsForPlatform,
-} from '@/components/kilo-chat/app-aware-keyboard-padding-state';
 import { type useNativeAuth } from '@/lib/auth/use-native-auth';
 import { useThemeColors } from '@/lib/hooks/use-theme-colors';
 import { canSubmitEmailCode } from './email-otp-state';
-
-const OTP_KEYBOARD_BREATHING_GAP = 16;
-
-function keyboardHeightFromEvent(event: KeyboardEvent): number {
-  return event.endCoordinates.height;
-}
 
 export function EmailOtpForm({
   email,
@@ -41,67 +23,11 @@ export function EmailOtpForm({
   const colors = useThemeColors();
   const codeRef = useRef('');
   const [hasCompleteCode, setHasCompleteCode] = useState(false);
-  const [keyboardHeight, setKeyboardHeight] = useState(0);
   const authBusy = busy !== undefined;
 
-  useEffect(() => {
-    const keyboardEvents = resolveKeyboardPaddingEventsForPlatform(Platform.OS);
-    if (keyboardEvents === null) {
-      setKeyboardHeight(0);
-      return undefined;
-    }
-
-    const keyboardShowSubscription = Keyboard.addListener(keyboardEvents.show, event => {
-      setKeyboardHeight(current =>
-        resolveAppAwareKeyboardPadding({
-          currentPadding: current,
-          event: {
-            type: 'keyboard-visible',
-            keyboardHeight: keyboardHeightFromEvent(event),
-          },
-        })
-      );
-    });
-    const keyboardHideSubscription = Keyboard.addListener(keyboardEvents.hide, () => {
-      setKeyboardHeight(current =>
-        resolveAppAwareKeyboardPadding({
-          currentPadding: current,
-          event: { type: 'keyboard-hidden' },
-        })
-      );
-    });
-    const appStateSubscription = AppState.addEventListener('change', appState => {
-      setKeyboardHeight(current =>
-        resolveAppAwareKeyboardPadding({
-          currentPadding: current,
-          event: { type: 'app-state-change', appState },
-        })
-      );
-    });
-
-    return () => {
-      keyboardShowSubscription.remove();
-      keyboardHideSubscription.remove();
-      appStateSubscription.remove();
-    };
-  }, []);
-
-  // Only pad while the keyboard is up, so the resting (keyboard-hidden) layout
-  // is not shifted within the parent's justify-center container.
-  const bottomSpacer = keyboardHeight > 0 ? keyboardHeight + OTP_KEYBOARD_BREATHING_GAP : 0;
-
   return (
-    // Defect A / QB-16: on iOS at Dynamic Type XXXL, the number pad occludes
-    // the Verify / Resend / Back controls. We add a bottom spacer equal to the
-    // keyboard height (plus a small breathing gap) inside the scrollable
-    // content, so the parent ScrollView (with automaticallyAdjustKeyboardInsets
-    // on iOS, and the window-level behavior="height" KeyboardAvoidingView on
-    // Android) has enough scrollable room for the user to scroll the controls
-    // above the number pad. keyboardShouldPersistTaps="handled" on the parent
-    // lets the user tap the controls mid-scroll. We intentionally do NOT nest a
-    // KeyboardAvoidingView here — nested behavior="padding" KAVs compute ~0
-    // bottom padding and do not help when content already overflows.
-    <View className="gap-3" style={{ paddingBottom: bottomSpacer }}>
+    // Keyboard avoidance is owned by the login shell in login-screen.tsx.
+    <View className="gap-3">
       <Text variant="muted" className="text-center text-sm">
         Enter the code sent to {email}
       </Text>
